+++ hutool-http/src/test/java/com/xiaoleilu/hutool/http/test/HttpRequestTest.java
@@ -1,5 +1,7 @@
 package com.xiaoleilu.hutool.http.test;
 
+import java.io.File;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -10,16 +12,28 @@ import com.xiaoleilu.hutool.date.DateUtil;
 import com.xiaoleilu.hutool.date.TimeInterval;
 import com.xiaoleilu.hutool.http.HttpRequest;
 import com.xiaoleilu.hutool.http.HttpResponse;
+import com.xiaoleilu.hutool.http.HttpUtil;
+import com.xiaoleilu.hutool.http.ssl.SSLSocketFactoryBuilder;
+import com.xiaoleilu.hutool.io.FileUtil;
 import com.xiaoleilu.hutool.lang.Console;
+import com.xiaoleilu.hutool.util.CharsetUtil;
 
 /**
  * {@link HttpRequest}单元测试
+ * 
  * @author Looly
  *
  */
 public class HttpRequestTest {
 	final String url = "http://photo.qzone.qq.com/fcgi-bin/fcg_list_album?uin=88888&outstyle=2";
-	
+
+	@Test
+	@Ignore
+	public void getHttpsTest() {
+		String body = HttpRequest.get("https://www.gjifa.com/pc/").execute().body();
+		Console.log(body);
+	}
+
 	@Test
 	@Ignore
 	public void asyncHeadTest() {
@@ -28,7 +42,7 @@ public class HttpRequestTest {
 		Console.log(headers);
 		Console.log(response.body());
 	}
-	
+
 	@Test
 	@Ignore
 	public void asyncGetTest() {
@@ -40,7 +54,7 @@ public class HttpRequestTest {
 		long interval2 = timer.interval();
 		Console.log("Async response spend {}ms, body spend {}ms", interval, interval2);
 	}
-	
+
 	@Test
 	@Ignore
 	public void syncGetTest() {
@@ -51,5 +65,46 @@ public class HttpRequestTest {
 		Console.log(body.body());
 		long interval2 = timer.interval();
 		Console.log("Async response spend {}ms, body spend {}ms", interval, interval2);
+	}
+
+	@Test
+	@Ignore
+	public void customGetTest() {
+		// 自定义构建HTTP GET请求，发送Http GET请求，针对HTTPS安全加密，可以自定义SSL
+		HttpRequest request = HttpRequest.get(url)
+				// 自定义返回编码
+				.charset(CharsetUtil.CHARSET_GBK)
+				// 禁用缓存
+				.disableCache()
+				// 自定义SSL版本
+				.setSSLProtocol(SSLSocketFactoryBuilder.TLSv12);
+		Console.log(request.execute().body());
+	}
+
+	@Test
+	@Ignore
+	public void uploadTest() {
+		File file = FileUtil.file("D:\\face.jpg");
+
+		// 方法一：自定义构建表单
+		HttpRequest request = HttpRequest//
+				.post("http://localhost:8080/file/upload")//
+				.form("file", file)//
+				.form("fileType", "图片");
+		HttpResponse response = request.execute();
+		System.out.println(response.body());
+	}
+
+	@Test
+	@Ignore
+	public void uploadTest2() {
+		File file = FileUtil.file("D:\\face.jpg");
+
+		// 方法二：使用统一的表单，Http模块会自动识别参数类型，并完成上传
+		HashMap<String, Object> paramMap = new HashMap<>();
+		paramMap.put("city", "北京");
+		paramMap.put("file", file);
+		String result = HttpUtil.post("http://wthrcdn.etouch.cn/weather_mini", paramMap);
+		System.out.println(result);
 	}
 }